pyproject.toml
[tool.poetry]
name = "web-scraping"
version = "0.1.0"
description = "Lab for web scraping"
authors = ["Silmara-Basso <silmarabasso@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
watermark = "^2.5.0"
selenium = "4.24.0"
beautifulsoup4 = "4.12.3"
pandas = "^2.2.3"
bs4 = "^0.0.2"
wordcloud = "^1.9.3"
pillow = "^10.4.0"
seaborn = "^0.13.2"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
